Complete rewrite of TREE core function, feels a bit saner now. Prefixes broken.
The new impl is inspired by depth-first-search as described in Artifical Intelligence: A Modern Approach by Russel and Norvig. Elements are placed in a LIFO list which is traversed recursively. The main recursive function (WALK-TREE) is now a toplevel function, so it is traceable. It also offers room for applying tail-call optimization - or it could be converted to a loop.
Right now, prefixes are not removed so every directory appends its prefixes and the output is wrong.