Split dir and file predicates, refactor *-dir-p
As a first experiment, PREDICATES has been changed to be a list of two
lists: directory predicates and file predicates. This means the
directory predicates get simpler but build-predicates and the actual
filtering are more complex.
This commit is rather an alpha state for experimenting and breaks the
tests for build-predicates, for example - I'll have to figure out how
to clean this up properly.

Nevertheless, 'filelimit' is now a valid parameter to cl-tree, e.g.:
./cl-tree --filelimit 40